+++ src/resources/repomap/queries/ruby-tags.scm
@@ -1,3 +1,7 @@
+; Ported from aider/queries/tree-sitter-language-pack/ruby-tags.scm at revision
+; 5dc9490bb35f9729ef2c95d00a19ccd30c26339c. Modified for Patch packaging.
+; Licensed under the Apache License, Version 2.0.
+
 ; Method definitions
 
 (